refactor(js/plugins/vertexai): rename context caching subdir
