+++ js/plugins/vertexai/src/context-caching/utils.ts
